+++ app/src/main/java/com/humolang/wifiless/data/model/Orientation.kt
@@ -1,7 +1,7 @@
 package com.humolang.wifiless.data.model
 
 data class Orientation(
-    val azimuth: Float = 0F,
-    val pitch: Float = 0F,
-    val roll: Float = 0F
+    val azimuth: Double = 0.0,
+    val pitch: Double = 0.0,
+    val roll: Double = 0.0
 )
